chess-corner - Copia/requirements.txt
flask
flask-cors
python-chess

